refactor: make reflowTablesAndDboItems be module
